Set license field of main paragraph to GPL-3+
"GPL-2+ and GPL-3+" is not possible for this field because it would basically
means that all those files can be distributed under GPL-2, which is
wrong (some files are GPL-3+ only).

The only acceptable license field for this paragraph is GPL-3+. Indeed we loose
the information that some of these files can also be distributed under the
GPL-2, but that's okay (the alternative would be to list individually all those
files, but it is too much work for a marginal benefit).

Git-Dch: Ignore